Beta 2.5.2 v1
Two features still need to be added:
1. Plugin dependent option removal still has a minor bug (you can call
it a feature) which will "delete" options from sight if any of the
following happens:
a. Robots options enabled, Robots.txt file has been added, Settings
Save, options are then returned disabled
b. The same happens for when you switch a doing it right theme to a
doing it wrong
c. And when you enable a sitemap plugin and then disable it

2. Sitemap filters.