Change date/time generation to use DATE_W3C format
The plugin is currently outputting the open graph article:published_time, article:modified_time and og:updated_time using only a Y-m-d format. This does not include any times, only the date.

We run a news site, and as important news breaks, it is common to have an article be modified one or more times during the day it was published. Rarely is the news article edited after the first day.

I have no idea how the Facebook scraper actually utilizes this tag, but it seems prudent to have a full date/time stamp for published and modified times so that their parser will know there has been an update.

This is a very simple change to use the PHP constant DATE_W3C.